comments, tidying up, etc
