+++ WhoSaidWhatNow/Utils/ChatUtils.cs
@@ -1,13 +1,43 @@
-using Dalamud.Logging;
 using ImGuiNET;
+using System;
 using System.Collections.Generic;
 using System.Numerics;
 using System.Text;
+using WhoSaidWhatNow.Objects;
 
 namespace WhoSaidWhatNow.Utils
 {
     public static class ChatUtils
     {
+        /// <summary>
+        /// Properly formats the passed data as a chat message and adds it to the log.
+        /// </summary>
+        public static void ShowMessage(KeyValuePair<DateTime, ChatEntry> c)
+        {
+            //tuple representing text that goes before/after player name
+            Tuple<string, string> tag = Plugin.Config.GUIFormats[c.Value.Type];
+
+            //dictionary for each chunk of text
+            Dictionary<string, Vector4> chunks = new Dictionary<string, Vector4>();
+
+            //timestamp. no intellisense i am not simplifying this, it looks like ass
+            chunks.Add($"[{c.Value.Time.ToShortTimeString()}] ", Plugin.Config.ChatColors[c.Value.Type]);
+
+            //if there isn't anything before the player name, ignore item1
+            if (tag.Item1 != String.Empty)
+                chunks.Add(tag.Item1, Plugin.Config.ChatColors[c.Value.Type]);
+
+            //leave out sender name if it's a standard emote
+            if (c.Value.Type != Dalamud.Game.Text.XivChatType.StandardEmote)
+                chunks.Add(c.Value.Sender.GetNameTag(), c.Value.Sender.NameColor);
+
+            //add message
+            chunks.Add(String.Format(tag.Item2, c.Value.Message), Plugin.Config.ChatColors[c.Value.Type]);
+
+            ChatUtils.WrappedColoredText(chunks);
+
+        }
+
         /// <summary>
         /// prints a single "paragraph" of colored text all on one line
         /// i never want to convert C++ code to C# ever again
@@ -66,7 +96,6 @@ namespace WhoSaidWhatNow.Utils
             //we printed the entire chat chunk, so put a manual newline
             ImGui.NewLine();
         }
-
         
         /// <summary>
         /// shorthand function for colored text
