+++ WhoSaidWhatNow/Utils/ConfigurationUtils.cs
@@ -1,7 +1,6 @@
 using Dalamud.Game.Config;
 using Dalamud.Game.Text;
 using Dalamud.Logging;
-using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Numerics;
