+++ WhoSaidWhatNow/Windows/TabGroups.cs
@@ -1,13 +1,11 @@
 using Dalamud.Interface;
 using ImGuiNET;
 using System;
-using System.Collections.Generic;
+using System.Linq;
 using System.Numerics;
 using WhoSaidWhatNow;
-using WhoSaidWhatNow.Windows;
-using System.Linq;
-using System.Text.RegularExpressions;
 using WhoSaidWhatNow.Utils;
+using WhoSaidWhatNow.Windows;
 
 public class TabGroups
 {
@@ -91,7 +89,7 @@ public class TabGroups
                                 var p = Plugin.Players.Find(p => c.Value.Sender.Name.Contains(p.Name));
                                 if (players[p!])
                                 {
-                                    MainWindow.ShowMessage(c);
+                                    ChatUtils.ShowMessage(c);
                                 }
                             }
                         }
